Renamed log.rs to log_directory.rs
Just to make it easier to navigate the code.
